+++ assets/js/admin/meta-boxes-product.js
@@ -20,7 +20,7 @@ jQuery( function( $ ) {
 	});
 
 	// Prevent enter submitting post form.
-	$( '#upsell_product_data' ).bind( 'keypress', function( e ) {
+	$( '#upsell_product_data' ).on( 'keypress', function( e ) {
 		if ( e.keyCode === 13 ) {
 			return false;
 		}
@@ -100,7 +100,7 @@ jQuery( function( $ ) {
 		var select_val = $( this ).val();
 
 		if ( 'variable' === select_val ) {
-			$( 'input#_manage_stock' ).change();
+			$( 'input#_manage_stock' ).trigger( 'change' );
 			$( 'input#_downloadable' ).prop( 'checked', false );
 			$( 'input#_virtual' ).removeAttr( 'checked' );
 		} else if ( 'grouped' === select_val ) {
@@ -113,11 +113,11 @@ jQuery( function( $ ) {
 
 		show_and_hide_panels();
 
-		$( 'ul.wc-tabs li:visible' ).eq( 0 ).find( 'a' ).click();
+		$( 'ul.wc-tabs li:visible' ).eq( 0 ).find( 'a' ).trigger( 'click' );
 
 		$( document.body ).trigger( 'woocommerce-product-type-change', select_val, $( this ) );
 
-	}).change();
+	}).trigger( 'change' );
 
 	$( 'input#_downloadable, input#_virtual' ).change( function() {
 		show_and_hide_panels();
@@ -165,7 +165,7 @@ jQuery( function( $ ) {
 
 		$( '.hide_if_' + product_type ).hide();
 
-		$( 'input#_manage_stock' ).change();
+		$( 'input#_manage_stock' ).trigger( 'change' );
 
 		// Hide empty panels/tabs after display.
 		$( '.woocommerce_options_panel' ).each( function() {
@@ -250,8 +250,8 @@ jQuery( function( $ ) {
 			$( 'p.stock_status_field:not( .hide_if_' + product_type + ' )' ).show();
 		}
 
-		$( 'input.variable_manage_stock' ).change();
-	}).change();
+		$( 'input.variable_manage_stock' ).trigger( 'change' );
+	}).trigger( 'change' );
 
 	// Date picker fields.
 	function date_picker_select( datepicker ) {
@@ -260,7 +260,7 @@ jQuery( function( $ ) {
 			date           = $( datepicker ).datepicker( 'getDate' );
 
 		$( otherDateField ).datepicker( 'option', option, date );
-		$( datepicker ).change();
+		$( datepicker ).trigger( 'change' );
 	}
 
 	$( '.sale_price_dates_fields' ).each( function() {
@@ -335,7 +335,7 @@ jQuery( function( $ ) {
 
 			attribute_row_indexes();
 
-			$attributes.find( '.woocommerce_attribute' ).last().find( 'h3' ).click();
+			$attributes.find( '.woocommerce_attribute' ).last().find( 'h3' ).trigger( 'click' );
 
 			$wrapper.unblock();
 
@@ -356,13 +356,13 @@ jQuery( function( $ ) {
 
 	$( '.product_attributes' ).on( 'click', 'button.select_all_attributes', function() {
 		$( this ).closest( 'td' ).find( 'select option' ).prop( 'selected', 'selected' );
-		$( this ).closest( 'td' ).find( 'select' ).change();
+		$( this ).closest( 'td' ).find( 'select' ).trigger( 'change' );
 		return false;
 	});
 
 	$( '.product_attributes' ).on( 'click', 'button.select_no_attributes', function() {
 		$( this ).closest( 'td' ).find( 'select option' ).removeAttr( 'selected' );
-		$( this ).closest( 'td' ).find( 'select' ).change();
+		$( this ).closest( 'td' ).find( 'select' ).trigger( 'change' );
 		return false;
 	});
 
@@ -436,7 +436,7 @@ jQuery( function( $ ) {
 					// Success.
 					$wrapper.find( 'select.attribute_values' )
 						.append( '<option value="' + response.term_id + '" selected="selected">' + response.name + '</option>' );
-					$wrapper.find( 'select.attribute_values' ).change();
+					$wrapper.find( 'select.attribute_values' ).trigger( 'change' );
 				}
 
 				$( '.product_attributes' ).unblock();
@@ -556,7 +556,7 @@ jQuery( function( $ ) {
 				}
 			});
 
-			file_path_field.val( file_path ).change();
+			file_path_field.val( file_path ).trigger( 'change' );
 		});
 
 		// Set post to 0 and set our custom type.