+++ tests/e2e/api/src/models/products/index.ts
@@ -1,0 +1,3 @@
+export * from './abstract-product';
+export * from './shared-types';
+export * from './simple-product';