+++ tests/e2e/env/utils/get-app-name.js
@@ -1,7 +1,7 @@
 /**
  * Provide the application name to bash scripts.
  */
-const getAppName = require( './app-name' );
+const { getAppName } = require( './app-name' );
 const appName = getAppName();
 
 console.log( appName );