+++ tests/e2e/core-tests/CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added new config variable for Simple Product price to `tests/e2e/env/config/default.json`. Defaults to 9.99
 - Merchant Product Edit tests
 - Merchant Product Search tests
+- Shopper Single Product tests
 - Shopper Checkout Apply Coupon